Create new standalone results page for DANDI uploads
