+++ src/renderer/src/stories/pages/uploads/UploadsPage.js
@@ -13,9 +13,15 @@ import { merge } from "../utils.js";
 import { run } from "../guided-mode/options/utils.js";
 import { notyf } from "../../../dependencies/globals.js";
 import Swal from "sweetalert2";
+import { Modal } from "../../Modal";
+import { DandiResults } from "../../DandiResults.js";
+
+export const isStaging = (id) => parseInt(id) >= 100000
 
 export async function uploadToDandi(info) {
-    if (!global.data.DANDI?.api_key) {
+
+    const api_key = global.data.DANDI?.api_key
+    if (!api_key) {
         await Swal.fire({
             title: "Your DANDI API key is not configured.",
             html: "Edit your settings to include this value.",
@@ -26,10 +32,11 @@ export async function uploadToDandi(info) {
         return this.to("settings");
     }
 
-    info.staging = parseInt(info.dandiset_id) >= 100000; // Automatically detect staging IDs
-    info.api_key = global.data.DANDI.api_key;
-
-    return await run("upload/folder", info, { title: "Uploading to DANDI" }).catch((e) => {
+    return await run("upload/folder", {
+        ...info, 
+        staging: isStaging(info.dandiset_id), // Automatically detect staging IDs
+        api_key
+    }, { title: "Uploading to DANDI" }).catch((e) => {
         this.notify(e.message, "error");
         throw e;
     });
@@ -48,15 +55,25 @@ export class UploadsPage extends Page {
             label: defaultButtonMessage,
             onClick: async () => {
                 await this.form.validate(); // Will throw an error in the callback
-                const results = await uploadToDandi.call(this, { ...global.data.uploads });
+                const results = await uploadToDandi.call(this, { ...globalState });
+
                 if (results)
                     notyf.open({
                         type: "success",
-                        message: `${global.data.uploads.nwb_folder_path} successfully uploaded to Dandiset ${global.data.uploads.dandiset_id}`,
+                        message: `${globalState.nwb_folder_path} successfully uploaded to Dandiset ${globalState.dandiset_id}`,
                     });
 
                 global.data.uploads = {};
                 global.save();
+
+                const modal = new Modal({ open: true })
+                modal.header = "DANDI Upload Summary"
+                const summary = new DandiResults({ id: globalState.dandiset_id })
+                summary.style.padding = '25px'
+                modal.append(summary)
+            
+                document.body.append(modal)
+
                 this.requestUpdate();
             },
         });
